Refactor code controller and ruby wasm init
