+++ app/javascript/ruby-wasm/index.js
@@ -1,0 +1,1 @@
+export * from './vm';
