+++ app/javascript/ruby-wasm/ruby/url.js
@@ -1,0 +1,2 @@
+const rubyWasmUrl =
+  'https://d21jaoini97l57.cloudfront.net/wasm-demo-app/0.0.2/ruby-3.3-web.wasm';
