--- app/javascript/ruby/index.js
@@ -1,3 +1,0 @@
-export * from './evaluator';
-export * from './rails';
-export * from './ruby';
